custom-elements/scoped-registry/ShadowRoot-createElement-duplicate-registrations.tentative.html

<!DOCTYPE html>
<meta name="author" title="Westbrook Johnson" href="mailto:westbrookjohnson@gmail.com">
<meta name="assert" content="Custom Elements should upgrade createElements without conflicts">
<link rel="help" href="https://wicg.github.io/webcomponents/proposals/Scoped-Custom-Element-Registries">
<link rel="help" href="https://github.com/WICG/webcomponents/issues/969">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id="testdiv"></div>

<script>
class TestAutonomousGlobal extends HTMLElement {};
customElements.define('test-element', TestAutonomousGlobal);

class TestAutonomousScoped extends HTMLElement {};

function attachShadowForTest(t, registry) {
  const host = document.createElement('div');
  const shadow = host.attachShadow({mode: 'open', registry});
  document.body.appendChild(host);
  t.add_cleanup(() => host.remove());
  return shadow;
}

test(t => {
  const registry = new CustomElementRegistry;
  registry.define('test-element', TestAutonomousScoped);

  const shadow = attachShadowForTest(t, registry);
  assert_true(shadow.createElement('test-element') instanceof TestAutonomousScoped);

}, 'Upgrade into autonomous custom element when inserted via innerHTML');
</script>
